+++ wallets/rn_cli_wallet_068_5/src/components/Modals/SignModal.tsx
@@ -2,40 +2,40 @@ import React from 'react';
 import {View, StyleSheet} from 'react-native';
 import Modal from 'react-native-modal';
 import {SignClientTypes} from '@walletconnect/types';
-import {Tag} from '../components/Tag';
-import {Methods} from '../components/Modal/Methods';
-import {Message} from '../components/Modal/Message';
-import {AcceptRejectButton} from '../components/AcceptRejectButton';
-import {ModalHeader} from '../components/Modal/ModalHeader';
+import {Tag} from '../Tag';
+import {Methods} from '../Modal/Methods';
+import {Message} from '../Modal/Message';
+import {getSignParamsMessage} from '../../utils/HelperUtils';
+import {AcceptRejectButton} from '../AcceptRejectButton';
+import {ModalHeader} from '../Modal/ModalHeader';
 import {
   approveEIP155Request,
   rejectEIP155Request,
-} from '../utils/EIP155Request';
-import {web3wallet} from '../utils/Web3WalletClient';
+} from '../../utils/EIP155Request';
+import {web3wallet} from '../../utils/Web3WalletClient';
 
-interface SendTransactionModalProps {
+interface SignModalProps {
   visible: boolean;
   setVisible: (arg0: boolean) => void;
   requestEvent: SignClientTypes.EventArguments['session_request'] | undefined;
   requestSession: any;
 }
 
-export function SendTransactionModal({
+export function SignModal({
   visible,
   setVisible,
   requestEvent,
   requestSession,
-}: SendTransactionModalProps) {
+}: SignModalProps) {
   const chainID = requestEvent?.params?.chainId?.toUpperCase();
   const method = requestEvent?.params?.request?.method;
+  const message = getSignParamsMessage(requestEvent?.params?.request?.params);
 
   const requestName = requestSession?.peer?.metadata?.name;
   const requestIcon = requestSession?.peer?.metadata?.icons[0];
   const requestURL = requestSession?.peer?.metadata?.url;
 
-  const {topic, params} = requestEvent;
-  const {request, chainId} = params;
-  const transaction = request.params[0];
+  const {topic} = requestEvent;
 
   async function onApprove() {
     if (requestEvent) {
@@ -71,7 +71,7 @@ export function SendTransactionModal({
             <Tag value={chainID} grey={true} />
           </View>
           <Methods methods={[method]} />
-          <Message message={JSON.stringify(transaction, null, 2)} />
+          <Message message={message} />
         </View>
 
         <View style={styles.flexRow}>
