+++ wallets/rn_cli_wallet_068_5/src/components/Modals/SignTypedDataModal.tsx
@@ -2,31 +2,31 @@ import React from 'react';
 import {View, StyleSheet} from 'react-native';
 import Modal from 'react-native-modal';
 import {SignClientTypes} from '@walletconnect/types';
-import {Tag} from '../components/Tag';
-import {Methods} from '../components/Modal/Methods';
-import {Message} from '../components/Modal/Message';
-import {getSignParamsMessage} from '../utils/HelperUtils';
-import {AcceptRejectButton} from '../components/AcceptRejectButton';
-import {ModalHeader} from '../components/Modal/ModalHeader';
+import {Tag} from '../Tag';
+import {Methods} from '../Modal/Methods';
+import {Message} from '../Modal/Message';
+import {getSignParamsMessage} from '../../utils/HelperUtils';
+import {AcceptRejectButton} from '../AcceptRejectButton';
+import {ModalHeader} from '../Modal/ModalHeader';
 import {
   approveEIP155Request,
   rejectEIP155Request,
-} from '../utils/EIP155Request';
-import {web3wallet} from '../utils/Web3WalletClient';
+} from '../../utils/EIP155Request';
+import {web3wallet} from '../../utils/Web3WalletClient';
 
-interface SignModalProps {
+interface SignTypedDataModalProps {
   visible: boolean;
   setVisible: (arg0: boolean) => void;
   requestEvent: SignClientTypes.EventArguments['session_request'] | undefined;
   requestSession: any;
 }
 
-export function SignModal({
+export function SignTypedDataModal({
   visible,
   setVisible,
   requestEvent,
   requestSession,
-}: SignModalProps) {
+}: SignTypedDataModalProps) {
   const chainID = requestEvent?.params?.chainId?.toUpperCase();
   const method = requestEvent?.params?.request?.method;
   const message = getSignParamsMessage(requestEvent?.params?.request?.params);
@@ -71,7 +71,7 @@ export function SignModal({
             <Tag value={chainID} grey={true} />
           </View>
           <Methods methods={[method]} />
-          <Message message={message} />
+          <Message message={JSON.stringify(message, null, 2)} />
         </View>
 
         <View style={styles.flexRow}>
